Split up the httphandlers package into multiple files
